requirements.txt
flask==3.0.2
python-dotenv==1.0.1
openai>=1.3.0
pytube==15.0.0
beautifulsoup4==4.12.3
PyPDF2==3.0.1
requests>=2.31.0
Markdown>=3.5.1
flask-wtf>=1.2.1
werkzeug>=2.3.7
pydantic==2.6.3
pydantic-settings==2.2.1
pytest>=7.4.3
pytest-cov>=4.1.0
mypy>=1.7.1
black>=23.11.0
flake8>=6.1.0
python-logging-loki>=0.3.1
structlog>=23.2.0
google-api-python-client==2.173.0
youtube-transcript-api==0.6.2
Pillow==10.2.0
pytesseract==0.3.10
gunicorn==21.2.0 